requirements.txt
adal==1.2.7
alabaster==1.0.0
albucore==0.0.21
albumentations==1.4.22
annotated-types==0.7.0
anyio==4.7.0
appnope==0.1.4
argcomplete==3.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.2.0
azure-ai-ml==1.22.4
azure-common==1.1.28
azure-core==1.32.0
azure-graphrbac==0.61.2
azure-identity==1.19.0
azure-mgmt-authorization==4.0.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-core==1.5.0
azure-mgmt-keyvault==10.3.1
azure-mgmt-network==28.0.0
azure-mgmt-resource==23.2.0
azure-mgmt-storage==21.2.1
azure-storage-blob==12.24.0
azure-storage-file-datalake==12.18.0
azure-storage-file-share==12.20.0
azureml-core==1.58.0.post1
babel==2.16.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bleach==6.2.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contextlib2==21.6.0
contourpy==1.3.1
coverage==7.6.9
cryptography==44.0.0
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.7.0
docker==7.1.0
docutils==0.21.2
email_validator==2.2.0
eval_type_backport==0.2.0
executing==2.1.0
fastapi==0.115.6
fastapi-cli==0.0.6
fastjsonschema==2.21.1
filelock==3.16.1
fonttools==4.55.3
fqdn==1.5.1
fsspec==2024.10.0
google-api-core==2.23.0
google-auth==2.36.0
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
humanfriendly==10.0
idna==3.10
imagesize==1.4.1
# Editable install with no version control (imagomum==0.1)
-e /Users/aceman/Documents/imagomum
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.30.0
ipywidgets==8.1.5
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.8.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpickle==3.4.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
knack==0.12.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
msrest==0.7.1
msrestazure==0.6.4.post1
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
ndg-httpsclient==0.5.1
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.1
notebook_shim==0.2.4
numpy==2.1.3
oauthlib==3.2.2
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
opencensus-ext-logging==0.1.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.12
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
paramiko==3.5.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
pkginfo==1.12.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
prometheus_client==0.21.1
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic-extra-types==2.10.1
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydash==8.0.4
Pygments==2.18.0
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==24.3.0
pyparsing==3.2.0
PySocks==1.7.1
pytest==8.3.4
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.12.0
rpds-py==0.22.3
rsa==4.9
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
SecretStorage==3.3.3
Send2Trash==1.8.3
setuptools==75.6.0
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==8.1.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
starlette==0.41.3
strictyaml==1.7.3
stringzilla==3.11.1
sympy==1.13.1
tabulate==0.9.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
torch==2.6.0.dev20241211
torchaudio==2.5.0.dev20241211
torchvision==0.20.0.dev20241211
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
widgetsnbextension==4.0.13
